Better check if opsiclientd directory exists.

+++ build_binary.sh
@@ -19,7 +19,7 @@
 cd src/
 python setup-cx-freeze.py build
 cd build/
-if [ -e opsiclientd ]; then
+if [ -e opsiclientd -a -d opsiclientd ]; then
 	echo "Removing old build directory."
 	rm -rf opsiclientd/
 fi
